Develop big merge 2

+++ src/data/brackets.js
@@ -92,6 +92,14 @@ const exportedMethods = {
       .toArray();
     return bracketList;
   },
+  async getBracketByTeamId(teamId) {
+    teamId = validation.checkId(teamId);
+    const bracketCollection = await brackets();
+    const bracketsWithTeam = await bracketCollection
+      .find({ teams: teamId })
+      .toArray();
+    return bracketsWithTeam;
+  },
   async createBracket(
     name,
     description,
@@ -120,7 +128,6 @@ const exportedMethods = {
     }
     const allPlayers = [];
     for (let i in teams) {
-      console.log(teams[i]);
       let team = await teamData.getTeamById(teams[i]);
       for (let j in team.playerIds) {
         allPlayers.push(team.playerIds[j]);
@@ -156,7 +163,9 @@ const exportedMethods = {
       _id: new ObjectId(bracketId),
     });
     if (!bracket) throw "Error: Bracket not found";
-
+    if (bracket.archive) {
+      throw "This bracket is archived and cannot be edited.";
+    }
     const updatedBracketData = {};
 
     if (updatedBracket.name) {
@@ -202,7 +211,6 @@ const exportedMethods = {
       try {
         await this.updateRecords(bracket.matches, bracket.teams);
       } catch (error) {
-        console.log(error);
         return;
       }
 
@@ -228,6 +236,9 @@ const exportedMethods = {
   async deleteBracket(id) {
     id = validation.checkId(id);
     const bracketCollection = await brackets();
+    const bracket = await this.getBracketById(id);
+    if (!bracket) throw "Error: Could not get bracket.";
+    await this.updateRecords(bracket.matches, bracket.teams);
     const deletionInfo = await bracketCollection.findOneAndDelete({
       _id: new ObjectId(id),
     });
@@ -376,8 +387,8 @@ const exportedMethods = {
     }
     for (let match of matches) {
       if (
-        (match.participants.length === 2) &
-        (match.participants[0].isWinner !== null)
+        match.participants.length === 2 &&
+        match.participants[0].isWinner !== null
       ) {
         for (let participant of match.participants) {
           log[participant.id].played += 1;
